+++ app/Http/Controllers/Reptiles/ReptileController.php
@@ -153,7 +153,7 @@ class ReptileController extends Controller
             'gender'    => ['required', 'max:1', 'in:M,F'],
             'birth'     => [ 'nullable'],
             'memo'      => [ 'string', 'nullable'],
-            'imgUrls'           => ['nullable', 'array'],
+            'imgUrls'           => ['nullable'],
         ];
         if($request->hasFile('images')){
             $validatedList['images'] = ['nullable', 'array'];
@@ -172,11 +172,18 @@ class ReptileController extends Controller
         $reqData = $validator->safe();
 
         $dbImgList = $reptile->img_urls;
-        $updateImgList = $reqData['imgUrls'];
+        $updateImgList = json_decode($reqData['imgUrls'], true);
+        
+        if($dbImgList === null){
+            $dbImgList = [];
+        }
+
         $deleteImgList = array_diff($dbImgList, $updateImgList);
 
-        $images = new ImageController();
-        $deleteResult = $images->deleteImages($deleteImgList);
+        if(!empty($deleteImgList)){
+            $images = new ImageController();
+            $deleteResult = $images->deleteImages($deleteImgList);
+        }
 
         if(gettype($deleteResult) !== 'boolean'){
             return $deleteResult;
